+++ packages/art/src/app/lib/conwayStepEngine.ts
